+++ CHANGELOG.md
@@ -28,6 +28,7 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 * Renamed `FieldDef` to `FieldDefinition` ([#92]), by [@localheinz]
 * Turned `$configuration` parameter of `FixtureFactory::defineEntity()` into `$afterCreate`, a `Closure` that will be invoked after object construction ([#101]), by [@localheinz]
 * Started throwing an `InvalidCount` exception instead of a generic `Exception` when an invalid number of entities are requested ([#105]), by [@localheinz]
+* Started throwing an `EntityDefinitionAlreadyRegistered` exception instead of a generic `Exception` when an entity definition for a class name has already been registered ([#106]), by [@localheinz]
 
 ### Fixed
 
@@ -51,5 +52,6 @@ For a full diff see [`fa9c564...master`][fa9c564...master].
 [#92]: https://github.com/ergebnis/factory-bot/pull/92
 [#101]: https://github.com/ergebnis/factory-bot/pull/101
 [#105]: https://github.com/ergebnis/factory-bot/pull/105
+[#106]: https://github.com/ergebnis/factory-bot/pull/106
 
 [@localheinz]: https://github.com/localheinz